Rename library ns's to grafter-2 & update RDF dependencies to RDF4j
grafter.tabular always had an implicit sesame dependency, so update
the library to work with RDF4j version of grafter.  And change
namespace to grafter-2 to allow projects to use both old and new
versions simultaneously should they require during a transition.

NOTE: this library is still deprecated and is soley in maintenance
mode.